Split oneDNN subgraph tests to multiple files
